+++ src/pages/hakjin/Main/Main.scss
@@ -304,6 +304,8 @@
 
   .story_profile {
     display: inline-block;
+
+    margin-top: 20px;
   }
 
   .other_story {
